Implémenté le memory manager pour gérer l'address space du PPC

+++ pefdump/CFM/FileMapping.h
@@ -15,7 +15,7 @@ namespace CFM
 {
 	class FileMapping
 	{
-		long fileSize;
+		long long fileSize;
 		void* address;
 		
 	public:
